requirements.txt
Flask
firebase-admin
flask_cors
together
python-dotenv